Added multine defs
Modified the value stored in env to also kep track of the value's metadata
These include whether the value is final and global, at this stage, but can later
include stuff like which sourcefile/line it was discovered in, etc.

The special variable $DERIVED is now called $__OUTPUT__.
Later, $__SELF__ and $__TARGET__ will be added to the mix.